backend/user_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1